services/django-rest-framework/generative_service/requirements.txt
absl-py==2.1.0
accelerate==0.33.0
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
astunparse==1.6.3
billiard==4.2.0
cachetools==5.5.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.7
djangorestframework==3.15.2
filelock==3.15.4
flash-attention==1.0.0
flatbuffers==24.3.7
fsspec==2024.6.1
gast==0.5.4
google-auth==2.38.0
google-genai==1.0.0
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
huggingface-hub==0.24.5
idna==3.7
install==1.3.5
Jinja2==3.1.2
keras==3.1.1
kombu==5.4.0
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.7
networkx==3.2.1
numpy==1.26.4
opt-einsum==3.3.0
optree==0.10.0
packaging==24.1
pika==1.3.2
pillow==10.2.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rsa==4.9
safetensors==0.4.2
six==1.16.0
sqlparse==0.5.1
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tokenizers==0.15.2
torch==2.2.1+cpu
torchaudio==2.2.1+cpu
torchvision==0.17.1+cpu
tqdm==4.66.4
transformers==4.39.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
uWSGI==2.0.26
vine==5.1.0
wcwidth==0.2.13
websockets==14.2
Werkzeug==3.0.1
wrapt==1.16.0
